C++
Sudoku Checker 
Once you have read and stored the puzzle, you should test each of the rules to make sure that the condition is satisfied (1-9 in each block, no two of the same value in the same column etc.).
Output
Your output should be Solution is good! if the solution satisfies all criterion and Wrong solution! if one of the rules is not followed.